+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `%` is the default pattern escape character instead of `#`.
 - More lenient number parsing (`001` is interpreted as `1`).
 
 ## [0.2.0] - 2021-02-14